apps/raspi/requirements.txt
RPi.GPIO
Seeed-grove.py
getmac
jsonschema
load_dotenv
pytz
referencing
requests
rpi_ws281x
schedule
